requirements.txt
Flask
scholarly
requests
python-dotenv